% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvh.R
\name{esc_kvh_v}
\alias{esc_kvh_v}
\title{Escape Special Characters in a value}
\usage{
esc_kvh_v(s)
}
\arguments{
\item{s}{string}
}
\value{
escaped string
}
\description{
Escape Newlines and Backslashes in a string which will be used as a key in a KVH file.
}
\details{
Escape is done by butting a backslash before a special character.'
}
