% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvh.R
\name{obj_by_keys}
\alias{obj_by_keys}
\title{Get Object Identified by its Keys.}
\usage{
obj_by_keys(li, keys)
}
\arguments{
\item{li}{a named nested list returned by kvh_read()}

\item{keys}{character vector naming key suites to identify an object}
}
\value{
an object corresponding to li[[keys[1]][[keys[2]][[...]]. Return
NULL if non valid keys.
}
\description{
Given a named nested list returned by kvh_read(), get a particular item from it.
The object is identified by a series of hierarchical keys, first key
corresponds to the first hierarchical level, the second corresponds to
the second and so on.
}
