% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kvh_read}
\alias{kvh_read}
\title{Parse file in KVH format}
\usage{
kvh_read(fn)
}
\arguments{
\item{fn}{character kvh file name.}
}
\description{
Returns a list with names formed form kvh keys and values formed from kvh values
If a kvh value has sub-keys, it is returned as a nested list. Otherwise it is
returned as a character string.
}
